requirements.txt
numpy
scipy
matplotlib
PIL
cv2
dlib
imutils
rasterio
pytorch
torchvision